little more work on kd-tree builder
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@115 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Parser.cpp
@@ -41,7 +41,7 @@ bool Parser::parseFloat(float& result_out)
 {
 	double x;
 	const bool result = parseDouble(x);
-	result_out = x;
+	result_out = (float)x;
 	return result;
